Add files via upload
piles_2 will create a static matplotlib chart of the percentile data for the past 2+ years of weekdays and plot days in March on top of it. AET_example2.png is an example. 

Notes from the file (also in the .py):

This will extract percentiles from the main data file for the AET and direction
specified in like 20(ish), it could certainly be looped. Right now it is not
efficient (it calculates the percentiles each time) and there are also several
other areas where the code could be tightened up and/or made more efficient.

Requires two inputs:
    
    AET must be defined 
    direct must be defined
    
    Approximately in line 30
    
One additional input, in line 90 or so, if you wish to use weekends instead of
weekdays, use >5 instead of <6.

Finally, the current chart shows the first 18 days of march, but the coloring 
could be changed to be made it more legible, this is mostly a proof of concept.